Moved PhiNet convolutional block in networks
