+++ docs/source/conf.py
@@ -27,7 +27,7 @@ extensions = [
 templates_path = ['_templates']
 exclude_patterns = []
 
-
+autoclass_content = 'both'
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
